Added main files
I will start my react project from here. First out, I will do everything in this project based off of the Desafio Ápice Sistemas.pdf file that was given to me to start off this project and maybe make it look a little better
than how it is detailed, as the instructions were merely what should be in the system and not how they should be done inch by inch. Thus, I have added the main App.js file with the Header component, that currently only has
the ADMIN text and line. After, it has the dropdowns requested. Then, on the footer of the page, it has the local, plus the date and time of Brasilia, as per request on the document.