Added ListaDePessoas.js and ListaDePessoas.css
Now, an admin can search for someone as well, not just register them on the interface. I also did some modifications on the Cadastro part so that it becomes a little bit cleaner to register someone. Another thing I look forward to is the Relatório and the Filtro de vendas part, where I will need to show the Pessoa/person (or venda aka sale) that the admin searched for